requirements.txt
username = victor
password  = Password2021
email = emekadefirst@gmail.com
django-admin-material-dashboard